requirements.txt
django
pillow
django-organizations
whitenoise
bcrypt
django-friendship